add step-by-step generation tab

+++ src/common.py
@@ -10,7 +10,7 @@ SONGS_DIR = os.path.join(BASE_DIR, "songs")
 TEMP_AUDIO_DIR = os.path.join(SONGS_DIR, "temp")
 
 
-def display_progress(message, percent, progress=None):
+def display_progress(message, percent=None, progress=None):
     if progress is None:
         print(message)
     else:
@@ -28,6 +28,8 @@ def remove_suffix_after(text: str, occurrence: str):
 def copy_files_to_new_folder(file_paths, folder_path):
     os.makedirs(folder_path)
     for file_path in file_paths:
+        if not os.path.exists(file_path):
+            raise Exception(f"File not found: {file_path}")
         shutil.copyfile(
             file_path, os.path.join(folder_path, os.path.basename(file_path))
         )
@@ -55,6 +57,11 @@ def json_dump(thing, path):
         )
 
 
+def json_load(path, encoding=None):
+    with open(path, encoding=encoding) as file:
+        return json.load(file)
+
+
 def get_hash(thing, size=5):
     return hashlib.blake2b(
         json_dumps(thing).encode("utf-8"), digest_size=size
@@ -73,6 +80,8 @@ def get_file_hash(filepath, size=5):
 def get_rvc_model(voice_model):
     rvc_model_filename, rvc_index_filename = None, None
     model_dir = os.path.join(RVC_MODELS_DIR, voice_model)
+    if not os.path.exists(model_dir):
+        raise Exception(f"Voice model directory '{voice_model}' does not exist.")
     for file in os.listdir(model_dir):
         ext = os.path.splitext(file)[1]
         if ext == ".pth":
